unfinished commit
app-emulation/wa-linux-agent: Since the contents of this package will
be now a part of the sysext image, we don't need any special
OEM-specific hacks. We don't need to install the package in
/usr/share/oem directory any more, so update the ebuild to use the
Gentoo python machinery to install files in the usual locations. This
can also use a normal python package, so replace dependencies on
dev-lang/python-oem and dev-python/distro-oem with dev-lang/python and
dev-python/distro, respectively. The waagent.conf file we used to
provide is updated (to disable auto updates, for example, and dropped
obsolete options) and now is a part of the patch, so it is installed
by the python machinery.

dev-lang/python-oem, dev-python/distro-oem: Drop the unnecessary
packages.

profiles, coreos-devel/board-packages: Drop things related to the
dropped python packages.

coreos-base/oem-azure: This package will be used for the sysext image,
instead of for installing files into /usr/share/oem. This means that
we can drop some files or move them elsewhere. The systemd service
file is not needed, because it is installed by the
app-emulation/wa-linux-agent package now. This also means that the
ignition file as lost its purpose. The grub.cfg and oem-release must
be installed in /usr/share/oem, next to the sysext raw image file, so
handling of these files is moved to the newly added
coreos-base/common-oem-files package. `eject` symlink to
`/usr/bin/true` is installed in the newly added manglefs.sh script.

coreos-base/common-oem-files: This is a new package that will generate
and install both grub.cfg and oem-release files into
/usr/share/oem. Each platform can customize the process by providing
their own fragments for each of the two files if necessary.